pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intelligent-trading-system"
version = "1.0.0"
description = "Intelligent Trading System with Anti-Loss Protection for SOL-BRL"
authors = [
    {name = "Everaldo Gomes"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "numpy>=2.3.0", 
    "scikit-learn>=1.7.0",
    "aiohttp>=3.12.0",
    "structlog>=25.4.0",
    "python-dotenv>=1.1.0",
    "mercado-bitcoin-python",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
notebook = [
    "ipykernel>=6.29.0",
    "notebook>=7.0.0",
    "jupyterlab>=4.0.0",
    "nest-asyncio>=1.6.0",  # For asyncio compatibility in notebooks
]

[project.scripts]
intelligent-trading = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "config/",
    "data/",
    "logs/",
    "*.md",
    "*.env.example"
]

[tool.uv]
package = true  # Enable entry points installation

[tool.uv.pip]
# Enable pip-like behavior for compatibility
index-strategy = "first-index"

[tool.uv.sources]
# Use local wheel for mercado_bitcoin_python
mercado-bitcoin-python = { path = "mercado_bitcoin_python-0.2.7-py3-none-any.whl" }
